% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_functions.R
\name{jobs.create}
\alias{jobs.create}
\title{Creates a job and returns it.}
\usage{
jobs.create(Job, onBehalfOfContentOwner = NULL)
}
\arguments{
\item{Job}{The \link{Job} object to pass to this method}

\item{onBehalfOfContentOwner}{The content owner's external ID on which behalf the user is acting on}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/yt-analytics-monetary.readonly
\item https://www.googleapis.com/auth/yt-analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/yt-analytics-monetary.readonly, https://www.googleapis.com/auth/yt-analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/reporting/v1/reports/}{Google Documentation}

Other Job functions: \code{\link{Job}}
}

