% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{ListJobsResponse}
\alias{ListJobsResponse}
\title{ListJobsResponse Object}
\usage{
ListJobsResponse(jobs = NULL, nextPageToken = NULL)
}
\arguments{
\item{jobs}{The list of jobs}

\item{nextPageToken}{A token to retrieve next page of results}
}
\value{
ListJobsResponse object
}
\description{
ListJobsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for ReportingService.ListJobs.
}

