% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_functions.R
\name{media.download}
\alias{media.download}
\title{Method for media download. Download is supportedon the URI `/v1/media/{+name}?alt=media`.}
\usage{
media.download(resourceName)
}
\arguments{
\item{resourceName}{Name of the media that is being downloaded}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/yt-analytics-monetary.readonly
\item https://www.googleapis.com/auth/yt-analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/yt-analytics-monetary.readonly, https://www.googleapis.com/auth/yt-analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/reporting/v1/reports/}{Google Documentation}
}
