% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{Media}
\alias{Media}
\title{Media Object}
\usage{
Media(resourceName = NULL)
}
\arguments{
\item{resourceName}{Name of the media resource}
}
\value{
Media object
}
\description{
Media Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Media resource.
}
