% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{Job}
\alias{Job}
\title{Job Object}
\usage{
Job(createTime = NULL, expireTime = NULL, reportTypeId = NULL,
  name = NULL, id = NULL, systemManaged = NULL)
}
\arguments{
\item{createTime}{The creation date/time of the job}

\item{expireTime}{The date/time when this job will expire/expired}

\item{reportTypeId}{The type of reports this job creates}

\item{name}{The name of the job (max}

\item{id}{The server-generated ID of the job (max}

\item{systemManaged}{True if this a system-managed job that cannot be modified by the user;}
}
\value{
Job object
}
\description{
Job Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A job creating reports of a specific type.
}
\seealso{
Other Job functions: \code{\link{jobs.create}}
}
