% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{ListReportsResponse}
\alias{ListReportsResponse}
\title{ListReportsResponse Object}
\usage{
ListReportsResponse(nextPageToken = NULL, reports = NULL)
}
\arguments{
\item{nextPageToken}{A token to retrieve next page of results}

\item{reports}{The list of report types}
}
\value{
ListReportsResponse object
}
\description{
ListReportsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for ReportingService.ListReports.
}
