% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{Media}
\alias{Media}
\title{YouTube Reporting API Objects 
Schedules reporting jobs containing your YouTube Analytics data and downloads the resulting bulk data reports in the form of CSV files.}
\usage{
Media(resourceName = NULL)
}
\arguments{
\item{resourceName}{Name of the media resource}
}
\value{
Media object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:23:23
filename: /Users/mark/dev/R/autoGoogleAPI/googleyoutubereportingv1.auto/R/youtubereporting_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Media Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Media resource.
}

