% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{Report}
\alias{Report}
\title{Report Object}
\usage{
Report(downloadUrl = NULL, startTime = NULL, createTime = NULL,
  jobId = NULL, id = NULL, endTime = NULL, jobExpireTime = NULL)
}
\arguments{
\item{downloadUrl}{The URL from which the report can be downloaded (max}

\item{startTime}{The start of the time period that the report instance covers}

\item{createTime}{The date/time when this report was created}

\item{jobId}{The ID of the job that created this report}

\item{id}{The server-generated ID of the report}

\item{endTime}{The end of the time period that the report instance covers}

\item{jobExpireTime}{The date/time when the job this report belongs to will expire/expired}
}
\value{
Report object
}
\description{
Report Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A report's metadata including the URL from which the report itself can bedownloaded.
}
