% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubereporting_objects.R
\name{ReportType}
\alias{ReportType}
\title{YouTube Reporting API Objects 
Schedules reporting jobs containing your YouTube Analytics data and downloads the resulting bulk data reports in the form of CSV files.}
\usage{
ReportType(id = NULL, systemManaged = NULL, deprecateTime = NULL,
  name = NULL)
}
\arguments{
\item{id}{The ID of the report type (max}

\item{systemManaged}{True if this a system-managed report type; otherwise false}

\item{deprecateTime}{The date/time when this report type was/will be deprecated}

\item{name}{The name of the report type (max}
}
\value{
ReportType object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:24:10
filename: /Users/mark/dev/R/autoGoogleAPI/googleyoutubereportingv1.auto/R/youtubereporting_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ReportType Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A report type.
}
